+++ Sources/PT.PM.Common/Json/UstJsonKeys.cs
@@ -1,0 +1,7 @@
+﻿namespace PT.PM.Common.Json
+{
+    public class UstJsonKeys
+    {
+        public const string KindName = "Kind";
+    }
+}
