+++ Sources/PT.PM.Matching/Json/PatternJsonConverterWriter.cs
@@ -1,0 +1,10 @@
+﻿using PT.PM.Common.Json;
+using System;
+
+namespace PT.PM.Matching.Json
+{
+    public class PatternJsonConverterWriter : UstJsonConverterWriter
+    {
+        public override bool CanConvert(Type objectType) => objectType.CanConvert();
+    }
+}
